refactor: organize into separate field components
